I created a relationship between Blog and Writer entities in Entity layer.
I added WriterLastBlog component in BlogReadAll page, the task of this component is to bring the blog posts according to the author id using the GetBlogListByWriter() method in BlogManager in the Business layer.

 I added the NewsLetter entity to the EntityLayer and created its table in the database with the Migration operation.
I created the necessary methods for NewsLetter entity in DataAccesLayer and BusinessLayer for adding messages.
I wrote SubscribeMail methods for the Partial that I will add to the BlogReadAll page in the NewsLetterController and created the SubscribeMail partial in the View file and added this partial to the BlogReadAll page.